Move anti-virus, export diagnostics, and available startup properties features out of premium
